PHP 8.5: replace non-canonical casts and the deprecated curl_close()
The base image moved from PHP 8.4.14 to 8.5.9, which deprecates the
non-canonical (integer) and (double) cast spellings and curl_close().

- ParameterHandler, Render, TemplateEngine, Translator: (integer) -> (int),
  (double) -> (float). Identical opcodes, purely the canonical spelling.
- Logger: drop curl_close(), a no-op since PHP 8.0 and deprecated in 8.5.
  The handle is freed when the last reference goes out of scope.
- CLAUDE.md: state the tested runtime (8.4+ floor per composer.json,
  tested on 8.5).
- .gitignore: ignore .codebase-memory/, matching the IdeaPlugin repo.

Verified on PHP 8.5.9: php -l with error_reporting=E_ALL reports no
deprecation over the whole tree, and tests/test_configLoader_INI.php
produces byte-identical output before and after the change.

+++ src/Eisodos/ParameterHandler.php
@@ -71,7 +71,7 @@
       
       // check REPOST
       if ($this->neq('PostID', '')) {
-        if ((integer)$this->getParam('PostID') <= (integer)$this->getParam('LastPostID')) {
+        if ((int)$this->getParam('PostID') <= (int)$this->getParam('LastPostID')) {
           $this->setParam('RePost', 'T', false, false, 'eisodos::parameterHandler');
         } else {
           $this->setParam('LastPostID', $this->getParam('PostID'), true, false, 'eisodos::parameterHandler');
@@ -79,7 +79,7 @@
       }
       
       // increase LastPostID
-      $this->setParam('PostID', (string)((integer)$this->getParam('LastPostID') + 1), false, false, 'eisodos::parameterHandler');
+      $this->setParam('PostID', (string)((int)$this->getParam('LastPostID') + 1), false, false, 'eisodos::parameterHandler');
       
       if ($this->neq('Reload', 'F')) {
         if ((string)$crc === $this->getParam('CRC')) {
@@ -365,7 +365,7 @@
      * Loads session variables into the parameter array and filters them by the rules defined in the .params file
      */
     private function _loadSessionVariables(): void {
-      mt_srand((double)microtime() * 1000000);
+      mt_srand((float)microtime() * 1000000);
       if (!session_id()) {
         session_set_cookie_params($this->getCookieParams());
         session_name($this->getParam('_environment') . ($this->getParam('_environment') ? '-' : '') . Eisodos::$applicationName);
